refactor(plugin): extract gpu_context/scope_token.rs (#1092)
Lift with_full_scope_or_err — the cdylib-issued scope-token bridge
every FullAccess vtable callback uses to validate its gpu_handle
slot — out of gpu_context/mod.rs into a dedicated submodule.
Documents the post-C3 escalate flow in one place: cdylib calls
escalate_begin to mint a u64 token bound to an Arc<GpuContext>,
hands the token back as every FullAccess method's gpu_handle, and
each callback resolves the token via this helper before
dispatching.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
